refactor(rust): modularize Archetype kernel
