+++ crates/source-runtime-next/src/archetype/tests/vulnerability.rs
@@ -1,0 +1,30 @@
+use super::*;
+
+#[test]
+fn vulnerability_identity_is_scoped_to_the_request_scan() {
+    let kernel = kernel(ArchetypeFamily::Vulnerability);
+    let scan = scan(&kernel);
+    let request = kernel.plan_vulnerabilities(scan.id).unwrap();
+    let records = kernel
+            .decode_vulnerabilities(
+                &request,
+                br#"[{"id":1,"scan_id":9,"line_number":42,"file_path":"app/main.py","category":"ssrf","severity":"high","created_at":"2026-08-19T18:03:00-07:00"}]"#,
+                &scan,
+                None,
+            )
+            .unwrap();
+    assert_eq!(records[0].provider_id, "archetype-vulnerability-9-1");
+    assert_eq!(
+        records[0].fields.get("line_number").map(String::as_str),
+        Some("42")
+    );
+    assert_eq!(
+        records[0].occurred_at.as_deref(),
+        Some("2026-08-20T01:03:00Z")
+    );
+
+    let error = kernel
+        .decode_vulnerabilities(&request, br#"[{"id":1,"scan_id":10}]"#, &scan, None)
+        .unwrap_err();
+    assert_eq!(error, ArchetypeError::ResponseScopeMismatch);
+}
